refactor(backend): reorganize scripts and update deployment process
