+++ backend/start.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+node ./scripts/server && node ./scripts/dash &
